usecase/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.0.2
chromadb>=0.4.0
pypdf>=3.0.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
openai>=1.0.0
lxml>=4.9.0
html5lib>=1.1
requests>=2.31.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pydantic>=2.0.0